backend/requirements.txt
fastapi
uvicorn
requests
pydantic
sparqlwrapper
python-multipart
aiohttp
openai
python-dotenv 